--- src/context/lexer/function-call.ts
@@ -1,6 +1,0 @@
-import { Primitive } from "./../object/primitive";
-
-export type FunctionCall = ({
-  identifier: string;
-  parameters: (FunctionCall | Primitive)[];
-});
