--- src/context/object/primitive.ts
@@ -1,2 +1,0 @@
-
-export type Primitive = (boolean | number | string);
