+++ src/lang/primitive.ts
@@ -1,0 +1,7 @@
+
+/**
+ * A `Primitive` represents a fundamental data type that can be represented in
+ * the SignText language, of which there are three: `boolean`, `number`, and
+ * `string`.
+ */
+export type Primitive = (boolean | number | string);
