+++ src/lang/token/function-call.ts
@@ -1,0 +1,8 @@
+import { Primitive } from "./../primitive";
+
+export type FunctionParameter = (FunctionCall | Primitive);
+
+export type FunctionCall = ({
+  identifier: string;
+  parameters: FunctionParameter[];
+});
